VBHXHMOB-168

+++ lib/modules/declare_info_630b/data/model/declare_info_630b_response_data.dart
@@ -39,7 +39,7 @@ class DeclareInfo630bResponseData {
   final String? denNgay;
 
   /// Tổng số ngày
-  final int? tongSoNgay;
+  final double? tongSoNgay;
 
   /// Từ ngày đơn vị
   final String? tuNgayDonVi;
